+++ desktop/flipper-common/src/server-types.tsx
@@ -14,6 +14,7 @@ import {
   DeviceType,
   DownloadablePluginDetails,
   InstalledPluginDetails,
+  MarketplacePluginDetails,
   OS as PluginOS,
   UpdatablePluginDetails,
 } from './PluginDetails';
@@ -223,6 +224,10 @@ export type FlipperServerCommands = {
   ) => Promise<void>;
   'device-stop-screencapture': (serial: string) => Promise<string>; // file path
   'device-shell-exec': (serial: string, command: string) => Promise<string>;
+  'device-install-app': (
+    serial: string,
+    appBundlePath: string,
+  ) => Promise<void>;
   'device-forward-port': (
     serial: string,
     local: string,
@@ -248,6 +253,7 @@ export type FlipperServerCommands = {
   'keychain-read': (service: string) => Promise<string>;
   'keychain-unset': (service: string) => Promise<void>;
   'plugins-load-dynamic-plugins': () => Promise<InstalledPluginDetails[]>;
+  'plugins-load-marketplace-plugins': () => Promise<MarketplacePluginDetails[]>;
   'plugins-get-bundled-plugins': () => Promise<BundledPluginDetails[]>;
   'plugins-get-installed-plugins': () => Promise<InstalledPluginDetails[]>;
   'plugins-get-updatable-plugins': (
@@ -257,6 +263,9 @@ export type FlipperServerCommands = {
     plugin: DownloadablePluginDetails,
   ) => Promise<InstalledPluginDetails>;
   'plugin-source': (path: string) => Promise<string>;
+  'plugins-install-from-marketplace': (
+    name: string,
+  ) => Promise<InstalledPluginDetails>;
   'plugins-install-from-npm': (name: string) => Promise<InstalledPluginDetails>;
   'plugins-install-from-file': (
     path: string,