+++ desktop/flipper-plugin/src/ui/data-table/__tests__/DataTable.node.tsx
@@ -764,3 +764,273 @@ test('selection always has the latest state', () => {
 
   rendering.unmount();
 });
+
+test('open second panel and append', async () => {
+  const ds = createTestDataSource();
+  const ref = createRef<DataTableManager<Todo>>();
+  const rendering = render(
+    <DataTable
+      enableMultiPanels
+      dataSource={ds}
+      columns={columns}
+      tableManagerRef={ref}
+    />,
+  );
+  {
+    const elem = await rendering.findAllByText('test DataTable');
+    expect(elem.length).toBe(1);
+    expect(elem[0].parentElement?.parentElement).toMatchInlineSnapshot(`
+      <div
+        class="ant-dropdown-trigger css-1k3kr6b-TableBodyRowContainer e1luu51r1"
+      >
+        <div
+          class="css-1baxqcf-TableBodyColumnContainer e1luu51r0"
+          width="50%"
+        >
+          <span>
+            <span
+              style="background-color: rgb(255, 245, 102);"
+            />
+            test DataTable
+          </span>
+        </div>
+        <div
+          class="css-1baxqcf-TableBodyColumnContainer e1luu51r0"
+          width="50%"
+        >
+          <span>
+            <span
+              style="background-color: rgb(255, 245, 102);"
+            />
+            true
+          </span>
+        </div>
+      </div>
+    `);
+  }
+  // hide done
+  act(() => {
+    ref.current?.toggleSideBySide();
+  });
+  expect(Object.keys(ds.additionalViews).length).toBeGreaterThan(0);
+  act(() => {
+    ds.append({
+      title: 'Drink coffee',
+      done: false,
+    });
+  });
+  {
+    const elem = await rendering.findAllByText('Drink coffee');
+    expect(elem.length).toBe(2);
+  }
+  act(() => {
+    ds.append({
+      title: 'Drink tea',
+      done: false,
+    });
+  });
+  {
+    const elem = await rendering.findAllByText('Drink tea');
+    expect(elem.length).toBe(2);
+  }
+});
+
+test('open second panel and update', async () => {
+  const ds = createTestDataSource();
+  const ref = createRef<DataTableManager<Todo>>();
+  const rendering = render(
+    <DataTable
+      enableMultiPanels
+      dataSource={ds}
+      columns={columns}
+      tableManagerRef={ref}
+    />,
+  );
+  {
+    const elem = await rendering.findAllByText('test DataTable');
+    expect(elem.length).toBe(1);
+    expect(elem[0].parentElement?.parentElement).toMatchInlineSnapshot(`
+      <div
+        class="ant-dropdown-trigger css-1k3kr6b-TableBodyRowContainer e1luu51r1"
+      >
+        <div
+          class="css-1baxqcf-TableBodyColumnContainer e1luu51r0"
+          width="50%"
+        >
+          <span>
+            <span
+              style="background-color: rgb(255, 245, 102);"
+            />
+            test DataTable
+          </span>
+        </div>
+        <div
+          class="css-1baxqcf-TableBodyColumnContainer e1luu51r0"
+          width="50%"
+        >
+          <span>
+            <span
+              style="background-color: rgb(255, 245, 102);"
+            />
+            true
+          </span>
+        </div>
+      </div>
+    `);
+  }
+  // hide done
+  act(() => {
+    ds.append({
+      title: 'Drink coffee',
+      done: false,
+    });
+  });
+  {
+    const elems = await rendering.findAllByText('Drink coffee');
+    expect(elems.length).toBe(1);
+  }
+  act(() => {
+    ref.current?.toggleSideBySide();
+  });
+  expect(Object.keys(ds.additionalViews).length).toBeGreaterThan(0);
+  {
+    const elems = await rendering.findAllByText('Drink coffee');
+    expect(elems.length).toBe(2);
+  }
+  act(() => {
+    ds.update(0, {
+      title: 'DataTable tested',
+      done: false,
+    });
+  });
+  {
+    const elems = await rendering.findAllByText('Drink coffee');
+    expect(elems.length).toBe(2);
+    expect(rendering.queryByText('test DataTable')).toBeNull();
+    const newElems = await rendering.findAllByText('DataTable tested');
+    expect(newElems.length).toBe(2);
+  }
+});
+
+test('open second panel and column visibility', async () => {
+  const ds = createTestDataSource();
+  const ref = createRef<DataTableManager<Todo>>();
+  const rendering = render(
+    <DataTable
+      enableMultiPanels
+      dataSource={ds}
+      columns={columns}
+      tableManagerRef={ref}
+    />,
+  );
+  {
+    const elem = await rendering.findAllByText('test DataTable');
+    expect(elem.length).toBe(1);
+    expect(elem[0].parentElement?.parentElement).toMatchInlineSnapshot(`
+      <div
+        class="ant-dropdown-trigger css-1k3kr6b-TableBodyRowContainer e1luu51r1"
+      >
+        <div
+          class="css-1baxqcf-TableBodyColumnContainer e1luu51r0"
+          width="50%"
+        >
+          <span>
+            <span
+              style="background-color: rgb(255, 245, 102);"
+            />
+            test DataTable
+          </span>
+        </div>
+        <div
+          class="css-1baxqcf-TableBodyColumnContainer e1luu51r0"
+          width="50%"
+        >
+          <span>
+            <span
+              style="background-color: rgb(255, 245, 102);"
+            />
+            true
+          </span>
+        </div>
+      </div>
+    `);
+  }
+
+  // toggle column visibility of first table(main panel)
+  act(() => {
+    ref.current?.toggleSideBySide();
+    ref.current?.toggleColumnVisibility('done');
+  });
+  {
+    const elem = await rendering.findAllByText('test DataTable');
+    expect(elem.length).toBe(2);
+    expect(elem[0].parentElement?.parentElement).toMatchInlineSnapshot(`
+      <div
+        class="ant-dropdown-trigger css-1k3kr6b-TableBodyRowContainer e1luu51r1"
+      >
+        <div
+          class="css-1baxqcf-TableBodyColumnContainer e1luu51r0"
+          width="50%"
+        >
+          <span>
+            <span
+              style="background-color: rgb(255, 245, 102);"
+            />
+            test DataTable
+          </span>
+        </div>
+      </div>
+    `);
+  }
+
+  act(() => {
+    ds.update(0, {
+      title: 'DataTable tested',
+      done: false,
+    });
+  });
+  {
+    expect(rendering.queryByText('test DataTable')).toBeNull();
+    const elem = await rendering.findAllByText('DataTable tested');
+    expect(elem.length).toBe(2);
+    expect(elem[0].parentElement?.parentElement).toMatchInlineSnapshot(`
+      <div
+        class="ant-dropdown-trigger css-1k3kr6b-TableBodyRowContainer e1luu51r1"
+      >
+        <div
+          class="css-1baxqcf-TableBodyColumnContainer e1luu51r0"
+          width="50%"
+        >
+          <span>
+            <span
+              style="background-color: rgb(255, 245, 102);"
+            />
+            DataTable tested
+          </span>
+        </div>
+      </div>
+    `);
+  }
+});
+
+test('open second panel and closing deletes dataView', async () => {
+  const ds = createTestDataSource();
+  const ref = createRef<DataTableManager<Todo>>();
+  render(
+    <DataTable
+      enableMultiPanels
+      dataSource={ds}
+      columns={columns}
+      tableManagerRef={ref}
+    />,
+  );
+  expect(Object.keys(ds.additionalViews).length).toBe(0);
+  act(() => {
+    ref.current?.toggleSideBySide();
+  });
+  expect(Object.keys(ds.additionalViews).length).toBe(1);
+  act(() => {
+    ref.current?.toggleSideBySide();
+  });
+  expect(Object.keys(ds.additionalViews).length).toBe(0);
+});